requirements.txt
# Core testing and data
pytest
pandas
requests
python-dotenv>=1.0.1

# Security
cryptography>=41.0.0

# File parsing
pymupdf
docx2txt
openpyxl
pypdf>=4.0
pygments

nltk
textstat
scikit-learn
numpy

# Code complexity and metrics
radon
lizard

# Media Processing
ffmpeg-python
opencv-python

# LLM
groq>=0.5.0

# Google Drive API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Export to docx and pdf
python-docx
reportlab
# pypdf>=4.0   # for testing (already listed in line 14)

# FastAPI and web server (for API endpoints)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Auth / Security
pyjwt>=2.8.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0

# Project thumbnails
pillow
matplotlib

# Add later, when the processing functions are built:

# scenedetect
# ultralytics
# openai-clip
# msal
